# platform/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_security.dcBody:
    properties:
      data_type:
        type: string
      field_name:
        type: string
      masking_strategy:
        type: string
      sensitivity_level:
        type: string
      table_ref:
        type: string
    required:
    - field_name
    - table_ref
    type: object
  internal_security.scanBody:
    properties:
      content:
        type: string
      scan_type:
        description: secret/sast/prompt/full（空=full）
        type: string
    required:
    - content
    type: object
...
info:
  contact: {}
  description: 企业 AI 原生研发平台后端（AI 驱动 需求→研发→测试→审批→发布 全流程）
  title: 智源 ANP 平台 API
  version: "1.0"
paths:
...
  /project-spaces/{id}/requirements/{rid}/generate-tests:
    post:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 需求ID
        in: path
        name: rid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 生成的测试用例
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 需求不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 根据需求生成测试用例
      tags:
      - qa
...
  /project-spaces/{id}/requirements/{rid}/run-tests:
    post:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 需求ID
        in: path
        name: rid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 批量验收结果(total/passed/failed/manual)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 批量运行需求的测试用例
      tags:
      - qa
  /project-spaces/{id}/security/audit-logs:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 审计日志列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 安全审计日志
      tags:
      - security
  /project-spaces/{id}/security/data-classifications:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 数据分级列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 数据分级列表
      tags:
      - security
    post:
      consumes:
      - application/json
      parameters:
      - in: path
        name: data_type
        type: string
      - in: path
        name: field_name
        required: true
        type: string
      - in: path
        name: masking_strategy
        type: string
      - in: path
        name: sensitivity_level
        type: string
      - in: path
        name: table_ref
        required: true
        type: string
      - description: 数据分级{field_name,table_ref,...}
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_security.dcBody'
      produces:
      - application/json
      responses:
        "200":
          description: 创建的数据分级
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 新增数据分级
      tags:
      - security
  /project-spaces/{id}/security/data-classifications/{did}:
    delete:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 数据分级ID
        in: path
        name: did
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除结果
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除数据分级
      tags:
      - security
  /project-spaces/{id}/security/findings:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 严重度过滤(critical/high/medium/low)
        in: query
        name: severity
        type: string
      - description: 状态过滤(open/suppressed)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 安全发现列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 安全发现列表
      tags:
      - security
  /project-spaces/{id}/security/findings/{fid}/suppress:
    post:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 发现ID
        in: path
        name: fid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 抑制结果(suppressed)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 抑制安全发现
      tags:
      - security
  /project-spaces/{id}/security/gate:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 闸门状态(风险汇总+决策)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 安全闸门
      tags:
      - security
  /project-spaces/{id}/security/scans:
    post:
      consumes:
      - application/json
      parameters:
      - in: path
        name: content
        required: true
        type: string
      - description: secret/sast/prompt/full（空=full）
        in: path
        name: scan_type
        type: string
      - description: 扫描请求{content,scan_type}
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_security.scanBody'
      produces:
      - application/json
      responses:
        "200":
          description: 扫描结果与发现{scan,findings}
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 触发安全扫描
      tags:
      - security
...
  /project-spaces/{id}/test-cases:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 测试用例列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 测试用例列表
      tags:
      - qa
  /project-spaces/{id}/test-cases/{tcid}/run:
    post:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 测试用例ID
        in: path
        name: tcid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 运行后的用例(含结果)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 用例不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 运行失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 运行单条测试用例(自动验收)
      tags:
      - qa
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"